+++ lib/core/enums/turbo_relation_type.dart
@@ -1,3 +1,4 @@
 enum TurboRelationType {
-    sourceTag,
+  sourceTag,
+  targetTag,
 }